Write version number in setup file
