+++ pyproject.toml
@@ -3,23 +3,3 @@
 [build-system]
 requires = ["setuptools>=61.0"]
 build-backend = "setuptools.build_meta"
-
-## While alternatives like Poetry and Flit define the package information in this file as well, Setuptools places this in a separate file.
-# [project]
-# name = "xlum-python"
-# version = "0.0.3"
-# authors = [
-#   { name="Grehl, Steve", email="research@steve-grehl.eu" },
-# ]
-# description = "python importer for the XLUM data exchange and archive format"
-# readme = "docs/PIP_README.md"
-# requires-python = ">=3.7"
-# classifiers = [
-#     "Programming Language :: Python :: 3",
-#     "License :: OSI Approved :: MIT License",
-#     "Operating System :: OS Independent",
-# ]
-
-# [project.urls]
-# "Homepage" = "https://github.com/SteveGrehl/xlum-python"
-# "Bug Tracker" = "https://github.com/SteveGrehl/xlum-python/issues"
