+++ setup.cfg
@@ -2,7 +2,7 @@
 # This is where all the configuration details of your package goes.
 [metadata]
 name = xlum
-version = attr: xlum.__version__
+version = 0.08
 author = Steve Grehl
 author_email = research@steve-grehl.eu
 url = https://github.com/SteveGrehl/xlum-python
